agentforge/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentforge"
version = "0.1.0"
description = "An open-source agent harness with pluggable memory strategies and built-in evaluation for benchmarking AI coding agents"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "MD Rabbi", email = "rabbim75@yahoo.com" }]
keywords = ["ai", "agents", "llm", "evaluation", "harness", "benchmark"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "tiktoken>=0.7.0",
    "rich>=13.0",
    "click>=8.0",
    "docker>=7.0",
    "pandas>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.5",
    "pre-commit>=3.0",
]
rag = [
    "chromadb>=0.5",
    "sentence-transformers>=3.0",
]
dashboard = [
    "fastapi>=0.110",
    "uvicorn>=0.30",
]

[project.scripts]
agentforge = "agentforge.cli:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
